[BACKLOG-17862] Support for simple literal equality predicates
like

   ( 1 = 0 ) OR ( 1 = 1 ) OR ( 'kitten' = 'dead' )

Added a check in SQLCondition to determine whether the
left and right elements of a condition clause appear to
be literals between an equals function.  If so treat
as literal eq check.

This is an ugly hack around our lack of a real parser.
Some kittens were killed in the making of this commit.

http://jira.pentaho.com/browse/BACKLOG-17862