[BACKLOG-18470] Equals in a HAVING clause mishandled
Added a check of the expression element to determine whether
it appears to be an aggregate expression, using the same criteria
from SQLField (it starts with one of the func names, followed by an
open param).

Redesigning dataservices to use a parser is the real way to
address these sorts of bugs, and will save the lives of countless
kittens.

http://jira.pentaho.com/browse/BACKLOG-18470